requirements.txt
asgiref==3.2.10
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.0.12
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
idna==3.4
iniconfig==1.1.1
packaging==21.3
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.11.0
PyJWT==2.1.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.20.0
pytz==2020.1
requests==2.26.0
sqlparse==0.3.1
toml==0.10.2
tomli==2.0.1
urllib3==1.26.12